pixelated zoom fixed, map was returning a float instead of a double

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: nkellum <nkellum@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/22 16:31:29 by nkellum           #+#    #+#             */
-/*   Updated: 2019/08/22 13:44:34 by nkellum          ###   ########.fr       */
+/*   Updated: 2019/08/22 16:36:59 by nkellum          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ typedef struct	s_mlx
 
 void			mandelbrot(t_mlx *mlx);
 void			plot(int x, int y, t_mlx *mlx, int iteration);
-float			map(double val, double start1, double stop1,
+double			map(double val, double start1, double stop1,
 	double start2, double stop2);
 void redraw(t_mlx *mlx);
 void zoom_in(t_mlx *mlx);
